refactor according to review

+++ debian-8-amd64/scripts/cleanup-virtualbox.sh
@@ -15,16 +15,18 @@
 #See the License for the specific language governing permissions and
 #limitations under the License.
 
-echo "Fill with 0 the swap partition to reduce box size"
-readonly swapuuid=$(/sbin/blkid -o value -l -s UUID -t TYPE=swap)
-readonly swappart=$(readlink -f /dev/disk/by-uuid/"$swapuuid")
-/sbin/swapoff "$swappart"
-dd if=/dev/zero of="$swappart" bs=1M || echo "dd exit code $? is ignored"
-/sbin/mkswap -U "$swapuuid" "$swappart"
+if test -f .vbox_version ; then
+    echo "Fill with 0 the swap partition to reduce box size"
+    readonly swapuuid=$(/sbin/blkid -o value -l -s UUID -t TYPE=swap)
+    readonly swappart=$(readlink -f /dev/disk/by-uuid/"$swapuuid")
+    /sbin/swapoff "$swappart"
+    dd if=/dev/zero of="$swappart" bs=1M || echo "dd exit code $? is ignored"
+    /sbin/mkswap -U "$swapuuid" "$swappart"
 
-echo "Fill filesystem with 0 to reduce box size"
-dd if=/dev/zero of=/EMPTY bs=1M || echo "dd exit code $? is ignored"
-rm -f /EMPTY
-# Block until the empty file has been removed, otherwise, Packer
-# will try to kill the box while the disk is still full and that's bad
-sync
+    echo "Fill filesystem with 0 to reduce box size"
+    dd if=/dev/zero of=/EMPTY bs=1M || echo "dd exit code $? is ignored"
+    rm -f /EMPTY
+    # Block until the empty file has been removed, otherwise, Packer
+    # will try to kill the box while the disk is still full and that's bad
+    sync
+fi
